requirements.txt
# API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# YouTube scraping
apify-client==1.12.2

# Google Sheets
google-auth-oauthlib==1.2.0
google-auth==2.27.0
gspread==5.12.0

# Utils
python-dotenv==1.0.0
